Fix iOS build (#260)
* wip remove webviews

* add test webview

* experiment with OpaqueBridge

* another mime type incorrect when trying to load the wasm file

* wip inlining wasm

* fix clear AsyncStorage on ios

* bring back the editor

* cleanup editor generation

* remove wasm asset parsing

* cleanup opaque

* add opaque index.html to gitignore

* fix web version

* cleanup gitignore

+++ apps/app/utils/authentication/clearLocalSessionData.ts
@@ -1,5 +1,10 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 
 export const clearLocalSessionData = async () => {
-  await AsyncStorage.clear();
+  // on iOS .clear would throw an exception in case there are no keys in the store
+  // https://github.com/react-native-async-storage/async-storage/issues/86#issuecomment-554257281
+  const asyncStorageKeys = await AsyncStorage.getAllKeys();
+  if (asyncStorageKeys.length > 0) {
+    AsyncStorage.clear();
+  }
 };
